requirements.txt
pymcprotocol>=0.2.0
pyyaml>=6.0.1 